adding paypage

+++ src/Heidelpay.Payment/Payment/BasketItem.cs
@@ -42,7 +42,7 @@ namespace Heidelpay.Payment
         /// Gets or sets the vat.
         /// </summary>
         /// <value>The vat.</value>
-        public int Vat { get; set; }
+        public double Vat { get; set; }
 
         /// <summary>
         /// Gets or sets the amount discount.
